BAU: Refactor Guice Module for S3ConfigSource
Extract the anonymous inner class used for binding `S3ConfigSource` into
its own class, simplifying `ConfigApplication`.
